arch: B6 refactor — split 4 large modules + add routes/notifications
Tackles the deferred B6 items from CODEC-FULL-AUDIT-2026-05-30.md.
Pytest: 2099 → 2128 (+29 regression tests, 0 regressions).

REFACTORS (4 phases, all back-compat via identity-equal re-exports)

P1 / SR-32: codec_hooks (1097 LOC → 879) split into runtime + trust
- New codec_plugin_trust.py (212 LOC) — SHA-256 allowlist data layer.
  Owns: _read_allowlist, _write_allowlist, _file_sha256,
  _is_plugin_allowed, _maybe_grandfather_existing_plugins,
  _emit_plugin_load_blocked, _ALLOWLIST_LOCK, _PLUGINS_DIR_DEFAULT,
  _PLUGINS_ALLOWLIST_DEFAULT, _PLUGIN_FILE_SUFFIX, path helpers.
- codec_hooks keeps the runtime — PluginRegistry, run_with_hooks,
  HookCtx/HookVeto, 5-event dispatcher, _fire_one_* helpers,
  _run_hook_with_timeout, _emit_hook_*, emit_operation_*,
  approve_plugin operator entry. Trust-store names re-imported at
  the top of the file so any external `codec_hooks._read_allowlist`
  etc. import keeps working.
- Test update: 3 of 4 plugin_load_blocked tests now monkeypatch
  codec_plugin_trust._log_event (the new emit home); the 4th
  (plugin_hook_timeout) stays on codec_hooks.

P2 / SR-33: codec_dashboard chat helpers → codec_chat_pipeline
- New codec_chat_pipeline.py (~170 LOC) owns _StepBudget, the 2
  config-reading helpers (_step_budget_enabled, _step_budget_for_route),
  and _is_conversational. Pure-data + side-effect-light — testable
  without standing up the FastAPI app.
- codec_dashboard re-imports for back-compat.
- chat_completion (608 LOC) STAYS in codec_dashboard — it threads too
  many implicit module-level state dependencies for a safe one-session
  move. This phase extracts the testable building blocks; the handler
  itself is a separate sprint.
- Test update: test_step_budget temp_config fixture now monkeypatches
  both codec_dashboard.CONFIG_PATH AND codec_chat_pipeline.CONFIG_PATH.

P3 / SR-34: notification endpoints → routes/notifications.py
- 4 endpoints moved: GET /api/notifications, GET /api/notifications/count,
  POST /api/notifications/{id}/read, POST /api/notifications/read-all.
- All deps already lived in routes/_shared (_notif_lock,
  _load_notifications, _write_notifications) — clean extraction.
- codec_dashboard now `app.include_router(notifications_router)`.
- Establishes the pattern: future groups (/api/qchat/*, /api/vibe/*,
  /api/heartbeat/*, /api/cortex/*, /api/audit/*, /api/observer/*)
  follow the same recipe one at a time.

P4 / SR-35: voice filters → codec_voice_filters.py
- New codec_voice_filters.py (96 LOC) owns NOISE_WORDS,
  WHISPER_HALLUCINATIONS, is_noise(), is_hallucination(), rms_int16().
  Stateless data + pure functions — tunable + testable without the
  WebSocket pipeline.
- codec_voice re-exports the constants for back-compat; _rms now
  delegates to rms_int16.
- VoicePipeline class slim down deferred — feed_audio reaches into
  too much per-pipeline state for a safe one-session extraction.

DEFERRED (still architecturally valuable, sprint-sized)
- chat_completion (608 LOC) FastAPI handler extraction
- VoicePipeline class split into VoiceVAD / VoiceTTS / VoiceSkillDispatch
- Remaining dashboard endpoint groups (/api/qchat, /api/vibe,
  /api/heartbeat, /api/schedules, /api/cortex, /api/audit, /api/observer,
  /api/prompts, /api/notifications got done in this PR)
- codec_agents crew-factory split into crews/ package

NEW TESTS

- tests/test_b6_refactor.py (29 tests, 4 classes):
  * TestB6P1HooksTrustSplit (6): identity-equal re-exports +
    end-to-end grandfather migration round-trip
  * TestB6P2ChatPipelineSplit (8): identity-equal re-exports +
    _is_conversational behavior matrix + _StepBudget consume/warn
  * TestB6P3NotificationsRoute (1): all 4 endpoints registered in app.routes
  * TestB6P4VoiceFiltersSplit (14): identity-equal re-exports +
    is_noise / is_hallucination / rms_int16 behavior matrix

Full pytest: 2128 passed, 83 skipped (env-dependent), 0 failed (1m38s).
Audit report: ~/codec-audit-reports/CODEC-FULL-AUDIT-2026-05-30.md
