+++ tests/test_step_budget.py
@@ -57,10 +57,16 @@ def temp_audit_log(tmp_path, monkeypatch):
 
 @pytest.fixture
 def temp_config(tmp_path, monkeypatch):
-    """Redirect codec_dashboard.CONFIG_PATH to a tmp file so we control what
-    _step_budget_for_route() reads on each call."""
+    """Redirect CONFIG_PATH to a tmp file so we control what
+    _step_budget_for_route() reads on each call.
+
+    B6-P2: _step_budget_for_route moved to codec_chat_pipeline; patch
+    there. Also patch codec_dashboard for any test that checks via the
+    re-export."""
     cfg = tmp_path / "config.json"
     monkeypatch.setattr(codec_dashboard, "CONFIG_PATH", str(cfg))
+    import codec_chat_pipeline
+    monkeypatch.setattr(codec_chat_pipeline, "CONFIG_PATH", str(cfg))
     return cfg
